backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sentence-transformers==5.2.0
tiktoken==0.12.0
beautifulsoup4==4.14.3
playwright==1.57.0
supabase==2.6.0
groq==0.13.1
youtube-transcript-api==1.2.3
pdfplumber==0.11.4
python-multipart==0.0.18
numpy
scikit-learn
scipy
torch
transformers
safetensors
tqdm
requests
reportlab
Pillow
